final version
final project version for presentation

+++ app/crud.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import Session
 import models, schemas
 import datetime
+import secrets
 from auth import hash_password
 
 def signup(username: str, password: str, db: Session):
@@ -23,23 +24,27 @@ def delete_user(user_id: int, db: Session):
 
 
 def create_ship(user_id: int, name: str, db: Session):
-    db_ship = models.Ship(user_id_refer = user_id,
-                          name = name)
-    
+    token = secrets.token_hex(32)
+
+    db_ship = models.Ship(
+        user_id_refer=user_id,
+        name=name,
+        token=token
+    )
+
     db.add(db_ship)
     db.commit()
     db.refresh(db_ship)
 
-
-def change_ship(ship_id: int, ship: schemas.ChangeShip, db: Session):
-    pass
+    return db_ship
 
 
 def delete_ship(ship_id: int, db: Session):
     db_ship = db.query(models.Ship).filter(models.Ship.ship_id == ship_id).first()
 
-    db.delete(db_ship)
-    db.commit()
+    if db_ship:
+        db.delete(db_ship)
+        db.commit()
 
 
 def post_data(ship_id_refer: int, data: str, timestamp: datetime, db: Session):
